fix various bugs in Modf, Add
Modf was not setting zero correctly in many cases. The Modf test
performs an addition, but addition of big.Ints used a path that
also modified one of the arguments. This has been fixed to use a new
tmp variable.

There are also various tests I found that are currently not working,
but I didn't want to dive deeper right now.