fix various bugs in Modf, Add
Modf was not setting zero correctly in many cases. The Modf test
performs an addition, but addition of big.Ints used a path that
also modified one of the arguments. This has been fixed to use a new
tmp variable.

There are also various tests I found that are currently not working,
but I didn't want to dive deeper right now.


Former-commit-id: 052c3c780c3e98514ded18411c9b4d7e9c8bcf1d [formerly 5fc6f19426581d356d95953ad5c72f052a40d87c]
Former-commit-id: 1963295ecc33c7e5ac02362da3e8af340431e862